fix warnings

+++ src/bixbyite_topaz.cpp
@@ -217,19 +217,19 @@ int main(int argc, char *argv[]) {
     constexpr int simd_size = 8;
     typedef Eigen::Matrix<float, simd_size, 3, Eigen::AutoAlign> SIMDVector3f;
 #pragma omp parallel for
-    for (size_t i = 0; i < events.rows() - simd_size; i += simd_size) {
+    for (int64_t i = 0; i < events.rows() - simd_size; i += simd_size) {
       SIMDVector3f vi, vf;
-      for (size_t j = 0; j < 3; ++j)
+      for (int j = 0; j < 3; ++j)
         vi.col(j) = events.block<simd_size, 1>(i, j + 5);
       for (const Eigen::Matrix3f &op : transforms2) {
         vf.transpose().noalias() = op * vi.transpose();
-        for (size_t j = 0; j < simd_size; ++j) {
+        for (int j = 0; j < simd_size; ++j) {
           h(vf(j, 0), vf(j, 1), vf(j, 2), weight(events(i + j, 0)));
         }
       }
     }
 #pragma omp parallel for collapse(2)
-    for (size_t i = events.rows() - events.rows() % simd_size; i < events.rows(); ++i) {
+    for (int64_t i = events.rows() - events.rows() % simd_size; i < events.rows(); ++i) {
       for (const Eigen::Matrix3f &op : transforms2) {
         Eigen::Vector3f vf = op * events.block<3, 1>(i, 5);
         h(vf[0], vf[1], vf[2], weight(events(i, 0)));
